--- BankApplication.Services/BankDatabase.cs
@@ -1,11 +1,0 @@
-﻿using BankApplication.Models;
-using System.Collections.Generic;
-
-
-namespace BankApplication.Services
-{
-    /*public class BankDatabase
-    {
-        public static List<Bank> Banks = new List<Bank>();
-    }*/
-}
